feat(sync): unify issue + PR sync into one cron-driven job + global header indicator
Sync was fragmented across four disconnected paths: an `issue-sync` cron
(issues only, open-only direct writes), an orphaned `prs-sync` cron that was
never scheduled (so PRs only ever synced on a manual button click — the root
cause behind the open/closed mismatch on /prs), and two separate manual buttons
(`syncAndDigest` for issues+digests+comments+PRs, `syncPullRequests` for PRs).
Issues and PRs drifted independently and the open-only cron couldn't correct
items closed upstream.

Collapse it into one path:

- Shared engine (`lib/sync/run-workspace-sync.ts`): `buildSyncContext` +
  `runSyncPhases` lift the 4 phases (issues → digests → comments → PRs) out of
  the inbox action so the cron and the manual trigger run identical logic, with
  the all-states fetch + `fullSyncedAt` marker applying to both.
- One cron (`/api/cron/sync`, every 5 min): `runSyncSweep` enqueues a deduped
  `sync` job per workspace (new `jobs.kind='sync'` + `jobs_sync_open_uniq`
  partial unique index, migration 0037); the existing dispatch worker claims it
  and runs `executeSyncJob` → the shared engine. Reuses the job queue's
  lease/retry/dedup so the LLM-bound digest phase survives serverless limits.
  Deletes the `issue-sync` and `prs-sync` routes; updates vercel.json +
  in-process scheduler (`CEZAR_CRON_SYNC_INTERVAL_MS`).
- Global header indicator (`components/sync-indicator.tsx` in TopBar): a small
  dot reflecting the active workspace's live `sync_status` over Realtime
  (idle/syncing/error/fresh), clickable to "sync now", with a hover tooltip
  showing last-sync time + status + counts. Replaces the three per-page sync
  buttons (removed from inbox/issues/prs; `prs-action.ts` + the orphaned
  `sync-button.tsx` deleted).

Typecheck green across @cezar/core and @cezar/gui.
