+++ .env.docker.example
@@ -82,8 +82,8 @@ CEZAR_USE_WORKFLOW_ENGINE=true
 # Vercel schedule (1 min / 10 min / 5 min).
 # CEZAR_DISPATCH_INTERVAL_MS=60000
 # CEZAR_TRIAGE_SWEEP_INTERVAL_MS=600000
-# CEZAR_CRON_ISSUE_SYNC_INTERVAL_MS=300000
-# CSV of cron paths to skip, e.g. /api/cron/issue-sync
+# CEZAR_CRON_SYNC_INTERVAL_MS=300000
+# CSV of cron paths to skip, e.g. /api/cron/sync
 # CEZAR_INPROCESS_CRON_DISABLED=
 # Base URL the in-process scheduler fetches its own cron routes on. Defaults to
 # the loopback http://127.0.0.1:${PORT:-3000} (the routes run in this same
