+++ packages/gui/.env.example
@@ -60,7 +60,7 @@ CEZAR_INPROCESS_CRON=
 CEZAR_INPROCESS_CRON_BASE_URL=
 
 # CSV of route paths to skip (optional), e.g.
-#   /api/cron/ci-fix,/api/cron/issue-sync
+#   /api/cron/ci-fix,/api/cron/sync
 CEZAR_INPROCESS_CRON_DISABLED=
 
 # Set to `false` to skip every legacy cron route (run only dispatch +
@@ -72,7 +72,7 @@ CEZAR_INPROCESS_CRON_LEGACY=
 # Defaults match `packages/gui/vercel.json` cadences. All optional.
 CEZAR_DISPATCH_INTERVAL_MS=60000              # 1 min
 CEZAR_TRIAGE_SWEEP_INTERVAL_MS=600000         # 10 min
-CEZAR_CRON_ISSUE_SYNC_INTERVAL_MS=300000      # 5 min
+CEZAR_CRON_SYNC_INTERVAL_MS=300000            # 5 min
 CEZAR_CRON_ISSUE_MATCH_INTERVAL_MS=60000      # 1 min
 CEZAR_CRON_ISSUE_FIX_INTERVAL_MS=60000        # 1 min
 CEZAR_CRON_CI_WATCH_INTERVAL_MS=60000         # 1 min
