+++ packages/gui/src/app/inbox/sync-action.ts
@@ -3,10 +3,8 @@
 import { getSessionUser } from '@/lib/auth';
 import { getActiveWorkspace } from '@/lib/workspace';
 import { createSupabaseAdminClient } from '@/lib/supabase/server';
-import { SupabaseStoreAdapter } from '@/lib/adapters/supabase-store';
-import { loadWorkspaceConfig } from '@/lib/load-workspace-config';
-import type { SupabaseClient } from '@supabase/supabase-js';
-import type { Database, SyncCounts, SyncPhase, SyncStatusState } from '@/lib/supabase/types';
+import { STALE_SYNC_MS, buildSyncContext, runSyncPhases, writeSyncStatus } from '@/lib/sync/run-workspace-sync';
+import type { Database } from '@/lib/supabase/types';
 
 // ─────────────────────────────────────────────────────────────────────
 // Sync & Digest — the global "pull from GitHub" action shared by the
@@ -22,11 +20,10 @@ import type { Database, SyncCounts, SyncPhase, SyncStatusState } from '@/lib/sup
 // refreshes when it lands on 'done'/'error'. The app is a long-lived Node
 // container (output: 'standalone', Docker/Dokploy), so a non-awaited promise
 // keeps running safely after the action returns.
+//
+// The actual phase pipeline lives in the shared `lib/sync/run-workspace-sync`
+// module so a cron/job worker can run the identical sync.
 // ─────────────────────────────────────────────────────────────────────
-
-/** A `syncing` row older than this is treated as stale (e.g. the container
- *  restarted mid-sync) and may be overwritten by a fresh sync. */
-const STALE_SYNC_MS = 10 * 60 * 1000;
 
 export interface SyncResult {
   /** True once the background sync has been kicked off (not when it finishes). */
@@ -35,25 +32,6 @@ export interface SyncResult {
 }
 
 type SyncStatusRow = Database['public']['Tables']['sync_status']['Row'];
-
-async function writeStatus(
-  supabase: SupabaseClient<Database>,
-  workspaceId: string,
-  patch: {
-    status: SyncStatusState;
-    phase?: SyncPhase | null;
-    message?: string | null;
-    counts?: SyncCounts;
-    error?: string | null;
-    started_at?: string | null;
-    finished_at?: string | null;
-  },
-): Promise<void> {
-  const { error } = await supabase
-    .from('sync_status')
-    .upsert({ workspace_id: workspaceId, ...patch }, { onConflict: 'workspace_id' });
-  if (error) console.warn('[syncAndDigest] sync_status write failed:', error.message);
-}
 
 export async function syncAndDigest(): Promise<SyncResult> {
   const user = await getSessionUser();
@@ -90,52 +68,22 @@ export async function syncAndDigest(): Promise<SyncResult> {
   }
   if (!token) return { ok: false, error: 'No GitHub token — sign out and back in to sync' };
 
-  const adapter = new SupabaseStoreAdapter(supabase, workspace.id);
-
-  // The store may be empty for a newly-connected workspace; tolerate that
-  // by seeding an empty store rather than failing the sync.
-  let store: Awaited<ReturnType<typeof core.IssueStore.fromPort>>;
+  let ctx: Awaited<ReturnType<typeof buildSyncContext>>;
   try {
-    store = await core.IssueStore.fromPort(adapter);
-  } catch {
-    store = await core.IssueStore.fromPort({
-      async load() {
-        return {
-          meta: {
-            owner: workspace.repoOwner,
-            repo: workspace.repoName,
-            lastSyncedAt: null,
-            fullSyncedAt: null,
-            totalFetched: 0,
-            version: 1 as const,
-            orgMembers: [],
-            orgMembersFetchedAt: null,
-          },
-          issues: [],
-        };
-      },
-      async save(data) {
-        await adapter.save(data);
-      },
-    });
-  }
-
-  let config: Awaited<ReturnType<typeof loadWorkspaceConfig>>;
-  try {
-    config = await loadWorkspaceConfig(workspace.id, supabase, {
-      githubToken: token,
+    ctx = await buildSyncContext({
+      supabase,
+      workspaceId: workspace.id,
       repoOwner: workspace.repoOwner,
       repoName: workspace.repoName,
+      token,
     });
   } catch (err) {
     return { ok: false, error: `Config load failed: ${err instanceof Error ? err.message : String(err)}` };
   }
 
-  const github = new core.GitHubService(config);
-
   // ── Flip the row to 'syncing' before we return, so the client sees the
   //    spinner immediately and the concurrency guard above is armed. ──
-  await writeStatus(supabase, workspace.id, {
+  await writeSyncStatus(supabase, workspace.id, {
     status: 'syncing',
     phase: 'issues',
     message: 'Fetching issues…',
@@ -146,10 +94,10 @@ export async function syncAndDigest(): Promise<SyncResult> {
   });
 
   // ── Run the four phases in the background; do NOT await. ──
-  void runSync({ supabase, workspaceId: workspace.id, store, github, config, llm: core.LLMService }).catch(
+  void runSyncPhases({ supabase, workspaceId: workspace.id, store: ctx.store, github: ctx.github, config: ctx.config }).catch(
     async (err) => {
       console.error('[syncAndDigest] background sync crashed:', err);
-      await writeStatus(supabase, workspace.id, {
+      await writeSyncStatus(supabase, workspace.id, {
         status: 'error',
         phase: null,
         error: err instanceof Error ? err.message : String(err),
@@ -159,170 +107,4 @@ export async function syncAndDigest(): Promise<SyncResult> {
   );
 
   return { ok: true };
-}
-
-interface RunSyncArgs {
-  supabase: SupabaseClient<Database>;
-  workspaceId: string;
-  store: Awaited<ReturnType<(typeof import('@cezar/core'))['IssueStore']['fromPort']>>;
-  github: InstanceType<(typeof import('@cezar/core'))['GitHubService']>;
-  config: Awaited<ReturnType<typeof loadWorkspaceConfig>>;
-  llm: (typeof import('@cezar/core'))['LLMService'];
-}
-
-/** The four serial sync phases, each writing its progress into `sync_status`.
- *  Phase 1 is fatal-on-error; phases 2–4 are best-effort (warn + continue)
- *  so a digest/comment/PR hiccup still produces a usable sync. */
-async function runSync({ supabase, workspaceId, store, github, config, llm: LLMService }: RunSyncArgs): Promise<void> {
-  const counts: SyncCounts = {};
-
-  // ── 1. Fetch issues ──
-  // Do a full (all-states, incl. closed) fetch until a complete sync has
-  // succeeded, then switch to incremental `since` fetches. The full fetch
-  // backfills closed issues — and corrects issues closed upstream — for stores
-  // first synced by the older open-only path, and bootstraps new workspaces.
-  try {
-    const meta = store.getMeta();
-    const fullSync = !meta.fullSyncedAt || !meta.lastSyncedAt;
-    const issues = fullSync
-      ? await github.fetchAllIssues(true)
-      : await github.fetchIssuesSince(meta.lastSyncedAt as string, true);
-    counts.issuesFetched = issues.length;
-    counts.issuesCreated = 0;
-    counts.issuesUpdated = 0;
-    for (const issue of issues) {
-      const r = store.upsertIssue(issue);
-      if (r.action === 'created') counts.issuesCreated += 1;
-      if (r.action === 'updated') counts.issuesUpdated += 1;
-    }
-    const nowIso = new Date().toISOString();
-    store.updateMeta({
-      lastSyncedAt: nowIso,
-      totalFetched: issues.length,
-      // Mark the store complete once the first all-states fetch lands, so
-      // later syncs go incremental.
-      ...(fullSync ? { fullSyncedAt: nowIso } : {}),
-    });
-    await store.save();
-  } catch (err) {
-    await writeStatus(supabase, workspaceId, {
-      status: 'error',
-      phase: null,
-      counts,
-      error: `Issue fetch failed: ${err instanceof Error ? err.message : String(err)}`,
-      finished_at: new Date().toISOString(),
-    });
-    return;
-  }
-
-  // ── 2. Generate digests for OPEN issues that don't have one yet ──
-  // Scoped to open issues: a full backfill can pull in hundreds of historical
-  // closed issues, and digesting those would be a large, low-value LLM spend.
-  try {
-    const needDigest = store.getIssues({ state: 'open', hasDigest: false });
-    if (needDigest.length > 0) {
-      await writeStatus(supabase, workspaceId, {
-        status: 'syncing',
-        phase: 'digests',
-        message: `Digesting ${needDigest.length} issue${needDigest.length === 1 ? '' : 's'}…`,
-        counts,
-      });
-      const service = new LLMService(config);
-      const issueData = needDigest.map((i) => ({ number: i.number, title: i.title, body: i.body }));
-      const results = await service.generateDigests(issueData, config.sync.digestBatchSize);
-      for (const [number, digest] of results) {
-        store.setDigest(number, digest);
-      }
-      counts.digestsCreated = results.size;
-      await store.save();
-    }
-  } catch (err) {
-    // Digest failure shouldn't abort the whole sync — comments + PR pull are
-    // still useful, and the user can retry to fill in the rest.
-    console.warn('[syncAndDigest] digest pass failed:', err);
-  }
-
-  // ── 3. Fetch comments for open issues that need them ──
-  try {
-    const needComments = store
-      .getIssues({ state: 'open' })
-      .filter((i) => !i.commentsFetchedAt && i.commentCount > 0);
-    if (needComments.length > 0) {
-      await writeStatus(supabase, workspaceId, {
-        status: 'syncing',
-        phase: 'comments',
-        message: `Fetching comments for ${needComments.length} issue${needComments.length === 1 ? '' : 's'}…`,
-        counts,
-      });
-      const commentMap = await github.fetchCommentsForIssues(needComments.map((i) => i.number));
-      for (const [num, comments] of commentMap) {
-        store.setComments(num, comments);
-      }
-      counts.commentsFetched = commentMap.size;
-      await store.save();
-    }
-  } catch (err) {
-    console.warn('[syncAndDigest] comments pass failed:', err);
-  }
-
-  // ── 4. Refresh PRs (all states) into the pull_requests table ──
-  try {
-    await writeStatus(supabase, workspaceId, {
-      status: 'syncing',
-      phase: 'prs',
-      message: 'Refreshing pull requests…',
-      counts,
-    });
-    // All states, newest-activity first, so PRs closed/merged upstream get
-    // their state corrected; cap the walk so a repo with thousands of historical
-    // PRs doesn't bloat this background pass.
-    const prs = await github.listPullRequests(500);
-    if (prs.length > 0) {
-      const rows = prs.map((p) => ({
-        workspace_id: workspaceId,
-        number: p.number,
-        title: p.title,
-        body: p.body,
-        state: p.state,
-        draft: p.draft,
-        labels: p.labels,
-        author: p.author,
-        html_url: p.htmlUrl,
-        head_sha: p.headSha,
-        head_ref: p.headRef,
-        base_ref: p.baseRef,
-        pr_created_at: p.createdAt,
-        pr_updated_at: p.updatedAt,
-      }));
-      const { error } = await supabase
-        .from('pull_requests')
-        .upsert(rows, { onConflict: 'workspace_id,number' });
-      if (!error) counts.prsUpdated = prs.length;
-    }
-  } catch (err) {
-    console.warn('[syncAndDigest] PR sync failed:', err);
-  }
-
-  // ── Done. ──
-  await writeStatus(supabase, workspaceId, {
-    status: 'done',
-    phase: null,
-    message: summarize(counts),
-    counts,
-    error: null,
-    finished_at: new Date().toISOString(),
-  });
-}
-
-function summarize(counts: SyncCounts): string {
-  const bits: string[] = [];
-  if (counts.issuesFetched) {
-    bits.push(
-      `${counts.issuesFetched} issue${counts.issuesFetched === 1 ? '' : 's'} (${counts.issuesCreated ?? 0} new · ${counts.issuesUpdated ?? 0} updated)`,
-    );
-  }
-  if (counts.digestsCreated) bits.push(`${counts.digestsCreated} digested`);
-  if (counts.commentsFetched) bits.push(`${counts.commentsFetched} commented`);
-  if (counts.prsUpdated) bits.push(`${counts.prsUpdated} PR${counts.prsUpdated === 1 ? '' : 's'}`);
-  return bits.length > 0 ? `Synced: ${bits.join(' · ')}` : 'Already up to date';
 }
