+++ packages/gui/src/lib/sync/resolve-workspace-token.ts
@@ -1,0 +1,29 @@
+import type { SupabaseClient } from '@supabase/supabase-js';
+import type { Database } from '@/lib/supabase/types';
+
+// Picks a workable GitHub token for the workspace: walk admins, return the
+// first one whose provider_token is stored. Falls back to the env-level
+// GITHUB_TOKEN so single-tenant self-hosted deployments still work.
+export async function resolveWorkspaceToken(
+  workspaceId: string,
+  supabase: SupabaseClient<Database>,
+): Promise<string | null> {
+  const { data: admins } = await supabase
+    .from('workspace_members')
+    .select('user_id')
+    .eq('workspace_id', workspaceId)
+    .eq('role', 'admin');
+
+  if (admins && admins.length > 0) {
+    const ids = admins.map((a) => a.user_id);
+    const { data: tokens } = await supabase
+      .from('user_github_tokens')
+      .select('provider_token')
+      .in('user_id', ids)
+      .limit(1);
+    const token = tokens?.[0]?.provider_token;
+    if (token) return token;
+  }
+
+  return process.env.GITHUB_TOKEN || null;
+}
